Structural refactoring.
