+++ cmd/kubectx/delete.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/pkg/errors"
 
-	"github.com/ahmetb/kubectx/internal/cmdutil"
 	"github.com/ahmetb/kubectx/internal/kubeconfig"
 	"github.com/ahmetb/kubectx/internal/printer"
 )
@@ -36,7 +35,7 @@ func (op DeleteOp) Run(_, stderr io.Writer) error {
 // deleteContext deletes a context entry by NAME or current-context
 // indicated by ".".
 func deleteContext(name string) (deleteName string, wasActiveContext bool, err error) {
-	kc := new(kubeconfig.Kubeconfig).WithLoader(cmdutil.DefaultLoader)
+	kc := new(kubeconfig.Kubeconfig).WithLoader(kubeconfig.DefaultLoader)
 	defer kc.Close()
 	if err := kc.Parse(); err != nil {
 		return deleteName, false, errors.Wrap(err, "kubeconfig error")
