+++ internal/kubeconfig/kubeconfigloader.go
@@ -1,27 +1,22 @@
-package cmdutil
+package kubeconfig
 
 import (
+	"github.com/ahmetb/kubectx/internal/cmdutil"
 	"os"
 	"path/filepath"
 
 	"github.com/pkg/errors"
-
-	"github.com/ahmetb/kubectx/internal/kubeconfig"
 )
 
 var (
-	DefaultLoader kubeconfig.Loader = new(StandardKubeconfigLoader)
+	DefaultLoader Loader = new(StandardKubeconfigLoader)
 )
 
 type StandardKubeconfigLoader struct{}
 
 type kubeconfigFile struct{ *os.File }
 
-func (*StandardKubeconfigLoader) Load() (kubeconfig.ReadWriteResetCloser, error) {
-	cfgPath, err := kubeconfigPath()
-	if err != nil {
-		return nil, errors.Wrap(err, "cannot determine kubeconfig path")
-	}
+func (*StandardKubeconfigLoader) Load(cfgPath string) (ReadWriteResetCloser, error) {
 	f, err := os.OpenFile(cfgPath, os.O_RDWR, 0)
 	if err != nil {
 		if os.IsNotExist(err) {
@@ -52,31 +47,9 @@ func kubeconfigPath() (string, error) {
 	}
 
 	// default path
-	home := HomeDir()
+	home := cmdutil.HomeDir()
 	if home == "" {
 		return "", errors.New("HOME or USERPROFILE environment variable not set")
 	}
 	return filepath.Join(home, ".kube", "config"), nil
-}
-
-func HomeDir() string {
-	if v := os.Getenv("XDG_CACHE_HOME"); v != "" {
-		return v
-	}
-	home := os.Getenv("HOME")
-	if home == "" {
-		home = os.Getenv("USERPROFILE") // windows
-	}
-	return home
-}
-
-// IsNotFoundErr determines if the underlying error is os.IsNotExist. Right now
-// errors from github.com/pkg/errors doesn't work with os.IsNotExist.
-func IsNotFoundErr(err error) bool {
-	for e := err; e != nil; e = errors.Unwrap(e) {
-		if os.IsNotExist(e) {
-			return true
-		}
-	}
-	return false
 }
