+++ internal/kubeconfig/kubeconfigloader_test.go
@@ -1,75 +1,14 @@
-package cmdutil
+package kubeconfig
 
 import (
+	"github.com/ahmetb/kubectx/internal/cmdutil"
 	"os"
 	"path/filepath"
 	"strings"
 	"testing"
 
-	"github.com/ahmetb/kubectx/internal/kubeconfig"
 	"github.com/ahmetb/kubectx/internal/testutil"
 )
-
-func Test_homeDir(t *testing.T) {
-	type env struct{ k, v string }
-	cases := []struct {
-		name string
-		envs []env
-		want string
-	}{
-		{
-			name: "XDG_CACHE_HOME precedence",
-			envs: []env{
-				{"XDG_CACHE_HOME", "xdg"},
-				{"HOME", "home"},
-			},
-			want: "xdg",
-		},
-		{
-			name: "HOME over USERPROFILE",
-			envs: []env{
-				{"HOME", "home"},
-				{"USERPROFILE", "up"},
-			},
-			want: "home",
-		},
-		{
-			name: "only USERPROFILE available",
-			envs: []env{
-				{"XDG_CACHE_HOME", ""},
-				{"HOME", ""},
-				{"USERPROFILE", "up"},
-			},
-			want: "up",
-		},
-		{
-			name: "none available",
-			envs: []env{
-				{"XDG_CACHE_HOME", ""},
-				{"HOME", ""},
-				{"USERPROFILE", ""},
-			},
-			want: "",
-		},
-	}
-
-	for _, c := range cases {
-		t.Run(c.name, func(tt *testing.T) {
-			var unsets []func()
-			for _, e := range c.envs {
-				unsets = append(unsets, testutil.WithEnvVar(e.k, e.v))
-			}
-
-			got := HomeDir()
-			if got != c.want {
-				t.Errorf("expected:%q got:%q", c.want, got)
-			}
-			for _, u := range unsets {
-				u()
-			}
-		})
-	}
-}
 
 func Test_kubeconfigPath(t *testing.T) {
 	defer testutil.WithEnvVar("HOME", "/x/y/z")()
@@ -119,12 +58,12 @@ func Test_kubeconfigPath_envOvverideDoesNotSupportPathSeparator(t *testing.T) {
 
 func TestStandardKubeconfigLoader_returnsNotFoundErr(t *testing.T) {
 	defer testutil.WithEnvVar("KUBECONFIG", "foo")()
-	kc := new(kubeconfig.Kubeconfig).WithLoader(DefaultLoader)
+	kc := new(Kubeconfig).WithLoader(DefaultLoader)
 	err := kc.Parse()
 	if err == nil {
 		t.Fatal("expected err")
 	}
-	if !IsNotFoundErr(err) {
+	if !cmdutil.IsNotFoundErr(err) {
 		t.Fatalf("expected ENOENT error; got=%v", err)
 	}
 }
